Refactor(tests): split addUser tests into positive and negative tests
